Fixed some warnings about type conversions. Fixed some comments.
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@990 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/X3DTexture2DNode.cpp
@@ -474,7 +474,7 @@ void X3DTexture2DNode::renderSubImage( Image *image, GLenum texture_target,
 
   unsigned char *modified_data = 
     new unsigned char[ width * height * bytes_per_pixel ]; 
-  for( unsigned int i = 0; i < height; i++ ) {
+  for( unsigned int i = 0; i < (unsigned int)height; i++ ) {
     memcpy( modified_data + i * width * bytes_per_pixel, 
             image_data + ( (i + y_offset )* image->width() + x_offset) * bytes_per_pixel,
             width * bytes_per_pixel );
